WIP: refactor phy.io.
