+++ phy/io/kwik/model.py
@@ -9,23 +9,22 @@
 import os.path as op
 from random import randint
 import os
-from collections import defaultdict
 
 import numpy as np
 
-from ..ext import six
-from .base_model import BaseModel
-from .h5 import open_h5, File
-from ..waveform.loader import WaveformLoader
-from ..waveform.filter import bandpass_filter, apply_filter
-from ..electrode.mea import MEA
-from ..utils.logging import debug, warn
-from ..utils.array import (PartialArray,
-                           _concatenate_virtual_arrays,
-                           _as_array,
-                           _unique,
-                           )
-from ..utils._types import _is_integer, _as_list, _is_list
+from ...ext import six
+from ..base_model import BaseModel, ClusterMetadata
+from ..h5 import open_h5, File
+from ...waveform.loader import WaveformLoader
+from ...waveform.filter import bandpass_filter, apply_filter
+from ...electrode.mea import MEA
+from ...utils.logging import debug, warn
+from ...utils.array import (PartialArray,
+                            _concatenate_virtual_arrays,
+                            _as_array,
+                            _unique,
+                            )
+from ...utils._types import _is_integer
 
 
 #------------------------------------------------------------------------------
@@ -257,88 +256,6 @@ class SpikeLoader(object):
     def __getitem__(self, item):
         times = self._spike_samples[item]
         return self._waveforms[times]
-
-
-#------------------------------------------------------------------------------
-# ClusterMetadata class
-#------------------------------------------------------------------------------
-
-class ClusterMetadata(object):
-    """Hold cluster metadata.
-
-    Features
-    --------
-
-    * New metadata fields can be easily registered
-    * Arbitrary functions can be used for default values
-
-    Notes
-    ----
-
-    If a metadata field `group` is registered, then two methods are
-    dynamically created:
-
-    * `group(cluster)` returns the group of a cluster, or the default value
-      if the cluster doesn't exist.
-    * `set_group(cluster, value)` sets a value for the `group` metadata field.
-
-    """
-    def __init__(self, data=None):
-        self._fields = {}
-        self._data = defaultdict(dict)
-        # Fill the existing values.
-        if data is not None:
-            self._data.update(data)
-
-    @property
-    def data(self):
-        return self._data
-
-    def _get_one(self, cluster, field):
-        """Return the field value for a cluster, or the default value if it
-        doesn't exist."""
-        if cluster in self._data:
-            if field in self._data[cluster]:
-                return self._data[cluster][field]
-            elif field in self._fields:
-                # Call the default field function.
-                return self._fields[field](cluster)
-            else:
-                return None
-        else:
-            if field in self._fields:
-                return self._fields[field](cluster)
-            else:
-                return None
-
-    def _get(self, clusters, field):
-        if _is_list(clusters):
-            return [self._get_one(cluster, field)
-                    for cluster in _as_list(clusters)]
-        else:
-            return self._get_one(clusters, field)
-
-    def _set_one(self, cluster, field, value):
-        """Set a field value for a cluster."""
-        self._data[cluster][field] = value
-
-    def _set(self, clusters, field, value):
-        clusters = _as_list(clusters)
-        for cluster in clusters:
-            self._set_one(cluster, field, value)
-
-    def default(self, func):
-        """Register a new metadata field with a function
-        returning the default value of a cluster."""
-        field = func.__name__
-        # Register the decorated function as the default field function.
-        self._fields[field] = func
-        # Create self.<field>(clusters).
-        setattr(self, field, lambda clusters: self._get(clusters, field))
-        # Create self.set_<field>(clusters, value).
-        setattr(self, 'set_{0:s}'.format(field),
-                lambda clusters, value: self._set(clusters, field, value))
-        return func
 
 
 #------------------------------------------------------------------------------
